move logging configuration to the api.py; move config loading to the api.py; keep logging_config file
