+++ scrapyd_k8s/__main__.py
@@ -1,8 +1,4 @@
-from .config_loader import config
-from .logging_config import setup_logging
+from .api import run
 
 if __name__ == "__main__":
-    logging_level = config.scrapyd().get('logging_level', 'INFO')
-    setup_logging(logging_level)
-    from .api import run  # Import after logging is configured
     run()
