--- scrapyd_k8s/config_loader.py
@@ -1,3 +1,0 @@
-from .config import Config
-
-config = Config()
